pyproject.toml
[project]
name = "code-recall"
version = "0.1.0"
description = "Persistent memory for coding agents via Mem0 + Qdrant + Ollama BGE-M3 + Gemini"
requires-python = ">=3.12"
dependencies = [
    "mem0ai>=0.1.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "ollama>=0.4.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["code_recall"]

[project.scripts]
code-recall-daemon = "code_recall.daemon:main"

[tool.ruff]
line-length = 120
